[add]添加products删除方法

+++ src/main/java/com/lushiqi/kinokocrm/controller/CustomersController.java
@@ -30,7 +30,7 @@ public class CustomersController {
     }
 
     @GetMapping("/delete/{id}")
-    public String deleteCustomer(@PathVariable Long id) {
+    public String deleteCustomer(@PathVariable Integer id) {
         repository.deleteById(id);
         return "redirect:/customers";
     }
